analysis/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
pandas==2.2.3
numpy==1.24.3
python-multipart==0.0.12
prophet==1.1.5
sqlalchemy==2.0.23
pymysql==1.1.0
aiomysql==0.2.0
cryptography==41.0.7
redis==5.0.1
boto3==1.34.0
plotly==5.18.0
httpx==0.25.2
python-jose[cryptography]==3.3.0
openai==1.55.3